Add files via upload
Added in excel writing capabilities and modified the excel file (not droppers yet, really--best to just copy paste in the ticker list) to be compatible with that. Created a sell loop that will cycle until all stocks are sold or the algorithm is manually stopped. Theoretically removes stocks from the column that they are analyzed in as the sell orders are generated to improve computation efficiency. Combined both buy and sell algorithms into one program that can theoretically be run as an efficient aid to manual trading. Buy and sell orders are currently given in the terminal. Future directions include either creating pop-up windows for manual trading or integrating with the schwab api for automated trading