Final Manual Algo
Haven't tested it live in the market, but it was able to generate the buy signals during market close hours, and then it appear to be going through the sell loop. Windows and Beeps all seem to check out. Does require user input as to whether you want to skip the buy cycle and how many trades you want to make