V3.Alpha--NOT FOR USE
Needs more stuff (make the stuff run based off of time), verify all of the requests outputs (especially regarding the settled cash balances) and that it can place buy and sell orders. Need to implement the optimized algo that is more selective. 

This is only for cash accounts. Margin account will be V4