Fixed the buy counter to where it cuts off
Needs to check over the df that stores the bought tickers and their quantities. Need to rework the sell section to ensure that it uses that df (no need to export it back out to the excel tbf) both for the tickers that it goes over and the quantities that it sells.