Attempted fix of Schwab API stuff + More (See descr.)
Added: 
config.yaml
Tickers-V3-Read.xlsx
V3.3-Cleaned.py
V3.3-Rough.py

Made a lot of changes. See code for details. Rough is the initial MFP, whereas with Cleaned I tried to remove as much of the unused stuff as I felt was advisable. I left in a few things because they might be handy for the next edits we need to do.

Completed:
-Schwab should now autorefresh tokens
-Removed share count limit, added $50 per position cap -- we need to go back in and put in the hard price controls (see to-do)
-Should now not trade with unsettled funds, added in buffer quantity that will be subtracted automatically from the settled funds amount -- can be modified to provide a control of how much of the account we invest.
-Removed how it was set up as a main() function -- if that needs to be put back in... well that will take two minutes
-Removed P/L estimation stuff -- can add back in and have it actually be functional later with Schwab API functionality

To-Do:
-Polygon API data vs schwab quote data for quality control **V. important
-Schwab API-based P/L estimation
-Optimization (for now at least) of the algo towards more short-term oriented trading to determine whether we have a valid alpha **V. important
-Put the hard price controls (only trade if price is between 1 & 50 dollars) earlier in the algorithm. **V. important

This list was intended to be extensive, but also I did a lot of stuff, so it sadly probably isn't. Will text or email with additional things if I remember anything